refactor(onboard): extract onboard-dashboard.ts
Move dashboard URL management, port forwarding, token retrieval,
access info, guidance lines, and the summary printer into a dedicated
module.

Functions needing runOpenshell/openshellShellCommand accept them as
DI parameters; onboard.ts provides thin wrappers. printDashboard
now uses getProviderLabel() from onboard-providers.
